+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Adding a column with a non-null default value is safe in Postgres 11+
 - Added check for `add_belongs_to`
 - Added check for `remove_columns`
+- Added check for `remove_reference`
 - Customized messages
 
 ## 0.2.3